Move client session objects to new module
